QSDK-8242 Provide Californium with the runtime - backport OSGI v4.2
