+++ californium-core/src/main/java/org/eclipse/californium/core/network/deduplication/CropRotation.java
@@ -30,6 +30,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.eclipse.californium.core.Utils;
+import org.eclipse.californium.core.network.deduplication.api.Deduplicator;
 import org.eclipse.californium.core.network.Exchange;
 import org.eclipse.californium.core.network.Exchange.KeyMID;
 import org.eclipse.californium.core.network.config.NetworkConfig;
