+++ element-connector/pom.xml
@@ -16,8 +16,8 @@
 	<description>Java socket abstraction for datagram transports (UDP, DTLS, etc.)</description>
 	
 	<properties>
-		<netty.version>4.1.1.Final</netty.version>
-		<netty.version.lowerbound>4.1</netty.version.lowerbound>
+		<netty.version>4.0.28.Final</netty.version>
+		<netty.version.lowerbound>4.0.28.Final</netty.version.lowerbound>
 		<netty.version.upperbound>5</netty.version.upperbound>
 	</properties>
 
@@ -53,7 +53,8 @@
 						<Export-Package>
 							org.eclipse.californium.elements,
 							org.eclipse.californium.elements.tcp,
-							org.eclipse.californium.elements.util
+							org.eclipse.californium.elements.util,
+							eu.javaspecialists.tjsn.concurrency.stripedexecutor
 						</Export-Package>
 						<Import-Package>
 							io.netty*; version="[${netty.version.lowerbound},${netty.version.upperbound})",
