+++ scandium-core/pom.xml
@@ -109,7 +109,8 @@
 							org.eclipse.californium.scandium.config,
 							org.eclipse.californium.scandium.dtls,
 							org.eclipse.californium.scandium.dtls.cipher,
-							org.eclipse.californium.scandium.dtls.pskstore
+							org.eclipse.californium.scandium.dtls.pskstore,
+							org.eclipse.californium.scandium.util.api
 						</Export-Package>
 						<Private-Package>
 							org.eclipse.californium.scandium.util
